chore: move forking code to its own directory (#609)
