+++ lib/forking/forked_blockchain.js
@@ -1,13 +1,13 @@
 var BlockchainDouble = require("../blockchain_double.js");
 var Account = require("ethereumjs-account").default;
 var Block = require("ethereumjs-block");
-var Log = require("./log.js");
-var Receipt = require("./receipt.js");
+var Log = require("../utils/log.js");
+var Receipt = require("../utils/receipt.js");
 var utils = require("ethereumjs-util");
-var ForkedStorageTrie = require("./forkedstoragetrie.js");
+var ForkedStorageTrie = require("./forked_storage_trie.js");
 var Web3 = require("web3");
-var to = require("./to.js");
-var Transaction = require("./transaction");
+var to = require("../utils/to.js");
+var Transaction = require("../utils/transaction");
 var async = require("async");
 var LRUCache = require("lru-cache");
 const BN = utils.BN;
@@ -121,9 +121,11 @@ function ForkedBlockchain(options) {
 
             pendingRequests.set(key, pendingRequest);
             // Node 8 doesn't have Promise.finally
-            promise.catch(() => {}).then(() => {
-              pendingRequests.delete(key);
-            });
+            promise
+              .catch(() => {})
+              .then(() => {
+                pendingRequests.delete(key);
+              });
             return promise;
           }
         })
