+++ lib/forking/forked_storage_trie.js
@@ -5,7 +5,7 @@ const checkpointInterface = require("merkle-patricia-tree/checkpoint-interface")
 var utils = require("ethereumjs-util");
 var inherits = require("util").inherits;
 var Web3 = require("web3");
-var to = require("./to.js");
+var to = require("../utils/to.js");
 
 inherits(ForkedStorageBaseTrie, BaseTrie);
