+++ lib/statemanager.js
@@ -8,7 +8,7 @@ var wallet = require("ethereumjs-wallet");
 var hdkey = require("ethereumjs-wallet/hdkey");
 var async = require("async");
 var BlockchainDouble = require("./blockchain_double.js");
-var ForkedBlockchain = require("./utils/forkedblockchain.js");
+var ForkedBlockchain = require("./forking/forked_blockchain.js");
 var Web3 = require("web3");
 var fs = require("fs");
 var sigUtil = require("eth-sig-util");
